flask_backend/requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-Cors==5.0.0
Flask-MySQL==1.6.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
PyMySQL==1.1.1
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
